Coverting to @exemplar/storybook-react
