+++ examples/.setup/addons.js
@@ -1,0 +1,5 @@
+try {
+  require('@storybook/addon-knobs/register');
+} catch (ex) {
+  console.dir(ex);
+}
