+++ examples/.setup/aliases.json
@@ -1,0 +1,3 @@
+{
+  "react-img-carousel": "../../src"
+}
