+++ examples/.setup/config.js
@@ -1,0 +1,4 @@
+import { addDecorator } from '@storybook/react';
+import { withKnobs } from '@storybook/addon-knobs';
+
+addDecorator(withKnobs);
