+++ examples/.setup/shared.scss
@@ -1,0 +1,2 @@
+@import '../../src/carousel.less';
+
